app/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
joblib==1.3.2
numpy==1.26.3
packaging==24.0
pickleshare==0.7.5
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.4.0
scipy==1.12.0
sqlparse==0.4.4
threadpoolctl==3.2.0
